login / signIn errors handling dao (finished + new methods) password encryption in DB (SHA-1)

+++ src/com/supinfo/suptrip/dao/TripDao.java
@@ -19,7 +19,7 @@ public interface TripDao {
     /**
      * Find a trip by id
      */
-    Trip findTripById(Long id);
+    Trip findTripById(int id);
 
     /**
      * @return an unmodifiable list of all trips stored in memory
